requirements.txt
streamlit
PyPDF2
python-docx
keybert
jobspy
pandas
